Functional Arthropod Traits

Data from: A summary of eight traits of Coleoptera, Hemiptera, Orthoptera and Araneae, occurring in grasslands in Germany.

Format

A data frame of 1230 observations and 17 variables.

Order

Suborder

Family

SpeciesID

Author

Body_Size

Mean body length (mm)

Dispersal_ability

ordinal scale of 0 = very low, 1 = low, 0,5 = medium, 0,75 = high, 1 = very high; Based on wing dimorphism, flying ability, activity ranges, dispersal strategies, individual movement and colonization dynamics, depending on taxon

Feeding_guild

Fine classification of feeding guild across larval and adult stages; less frequent assignments in brackets.

c

carnivore

c-d

carni-detritivore

c-d-h

carni-detriti-herbivore

c-f

carni-fungivore

c-h

carni-herbivore

c-(h)

mainly carnivore, rarely herbivore

d

detritivor

d-f

detriti-fungivore

d-h

detriti-herbivore

f

fungivore

f-h

fungi-herbivore

h

herbivor

h-(c)

mainly herbivore, rarely carnivore

Feeding_guild_short

Coarse classification of feeding guild, indicating main feeding source across larval and adult stages

Feeding_mode

The way nutrients are ingested

Feeding_specialization

Host plant specialization in herbivores

Feeding_tissue

Fine classification on the plant tissues sucking herbivores are feeding on

Feeding_plant_part

Fine classification on the plant parts chewing herbivores are feeding on

Endophagous_lifestyle

Details on endophagously living larvae

Stratum_use

Vertical strata used across larval and adult stages; less frequent assignments in brackets

Stratum_use_short

Main vertical stratum used across larval and adult stages

Remark

Indicates species that do neither obligatory nor facultative occur in grasslands; * = non grasland species

original description: https://www.nature.com/articles/sdata201513/tables/3

Source

doi:10.5061/dryad.53ds2 ; Creative Commons 0. To the extent possible under law, the authors have waived all copyright and related or neighboring rights to this data.

Value

This is a data object. provides instructions for pulldata().

Details

Analyses of species traits have increased our understanding of how environmental drivers such as disturbances affect the composition of arthropod communities and related processes. There are, however, few studies on which traits in the arthropod community are affected by environmental changes and which traits affect ecosystem functioning. The assembly of arthropod traits of several taxa is difficult because of the large number of species, limited availability of trait databases and differences in available traits. We sampled arthropod species data from a total of 150 managed grassland plots in three regions of Germany. These plots represent the spectrum from extensively used pastures to mown pastures to intensively managed and fertilized meadows. In this paper, we summarize information on body size, dispersal ability, feeding guild and specialization (within herbivores), feeding mode, feeding tissue (within herbivorous suckers), plant part (within herbivorous chewers), endophagous lifestyle (within herbivores), and vertical stratum use for 1,230 species of Coleoptera, Hemiptera (Heteroptera, Auchenorrhyncha), Orthoptera (Saltatoria: Ensifera, Caelifera), and Araneae, sampled by sweep-netting between 2008 and 2012. We compiled traits from various literature sources and complemented data from reliable internet sources and the authors’ experience.

The data set comprises literature trait data of species that were sampled and measured in a project within the Biodiversity Exploratories which focuses on the effect of land use on arthropod community composition and related processes (e.g. species interactions such as herbivory or predation) in three regions of Germany

When using this data, please cite the original publication:

Gossner MM, Simons NK, Achtziger R, Blick T, Dorow WHO, Dziock F, Köhler F, Rabitsch W, Weisser WW (2015) A summary of eight traits of Coleoptera, Hemiptera, Orthoptera and Araneae, occurring in grasslands in Germany. Scientific Data 2: 150013. doi:10.1038/sdata.2015.13

Additionally, please cite the Dryad data package:

Gossner MM, Simons NK, Achtziger R, Blick T, Dorow WHO, Dziock F, Köhler F, Rabitsch W, Weisser WW (2015) Data from: A summary of eight traits of Coleoptera, Hemiptera, Orthoptera and Araneae, occurring in grasslands in Germany. Dryad Digital Repository. doi:10.5061/dryad.53ds2

See also

Author

Gossner MM, Simons NK, Achtziger R, Blick T, Dorow WHO, Dziock F, Köhler F, Rabitsch W, Weisser WW