Carabid morphological traits

Average body measures of 120 Carabid species occuring in the Netherlands.

Format

A data frame containing following columns:

name_correct = species name

source_measurement = researcher who performed measurement;

body_length = body length in mm;

antenna_length = antenna length in mm;

metafemur_length = length metafemur in mm;

eyewidth_corr = eye width in mm;

note = note;

resid_femur = residual femur length in mm (i.e. residual from linear model in which femur length is explained by body length);

resid_eye = residual eye length in mm (i.e. residual from linear model in which eye length is explained by body length)

resid_antenna = residual antenna length in mm (i.e. residual from linear model in which antenna length is explained by body length)

Source

doi: 10.5061/dryad.53ds2 ; Creative Commons 0. To the extent possible under law, the authors have waived all copyright and related or neighboring rights to this data.

Details

When using this data, please cite the original publication:

van der Plas F, van Klink R, Manning P, Olff H, Fischer M (2017) Sensitivity of functional diversity metrics to sampling intensity. Methods in Ecology and Evolution 8(9): 1072-1080. doi: doi: 10.1111/2041-210x.12728

Additionally, please cite the Dryad data package:

van der Plas F, van Klink R, Manning P, Olff H, Fischer M (2017) Data from: Sensitivity of functional diversity metrics to sampling intensity. Dryad Digital Repository. doi: doi: 10.5061/dryad.1fn46

See also

Author

Fons van der Plas, R. van Klink, P. Manning, H. Olff, M. Fischer