BExIS access over a Webservice

Read table from a web service. Inherits functionality of read.table().

Usage

read.service(
  datasetid,
  user = NULL,
  pswd = NULL,
  dec = ".",
  na.strings = "NA",
  fill = FALSE,
  sep = "\t",
  quote = if (identical(sep, "\n")) "" else "'\"",
  fileEncoding = "UTF-8"
)

read.service.blocks(
  datasetid,
  user,
  pswd,
  dec = ".",
  na.strings = "NA",
  fill = FALSE,
  sep = "\t",
  quote = if (identical(sep, "\n")) "" else "'\"",
  fileEncoding = "UTF-8"
)

Arguments

datasetid

Integer BExIS ID of the requested dataset.

user

User name on BExIS. If not provided, function will prompt input.

pswd

Password on BExIS. If not provided, function will prompt input.

dec

the character used in the file for decimal points.

na.strings

a character vector of strings which are to be interpreted as NA values. Blank fields are also considered to be missing values in logical, integer, numeric and complex fields.

fill

logical. If TRUE then in case the rows have unequal length, blank fields are implicitly added. See 'Details' of ?read.table.

sep

the field separator character. Values on each line of the file are separated by this character. If sep = "" (the default for read.table) the separator is 'white space', that is one or more spaces, tabs, newlines or carriage returns.

quote

the set of quoting characters. To disable quoting altogether, use quote = "". See scan for the behaviour on quotes embedded in quotes. Quoting is only considered for columns read as character, which is all of them unless colClasses is specified.

fileEncoding

character string: if non-empty declares the encoding to be used on a file (not a connection) so the character data can be re-encoded as they are written. See base::connections().

Value

Returns dataset from BExIS as a data.frame.

Details

read.service.blocks() returns a data.frame array as a list of data.frames.

Author

Dennis Heimann, Andreas Ostrowski