Standardize scientific names of species

Adds columns to a traitdata object containing accepted species names and relates to globally unique taxon identifiers via URI.

standardize.taxonomy(x, method = get_gbif_taxonomy, infraspecies = FALSE,
  fuzzy = FALSE, verbose = TRUE, return = c("taxonID",
  "scientificNameStd", "order", "taxonRank"), ...)

Arguments

x
a traitdata object (as returned by as.traitdata()) or a data table containing at least the column `scientificName.

method
not functional. Will allow to chose from different sources of taxonomic reference.

infraspecies
not functional.

fuzzy
if set to FALSE (default mode), this disables fuzzy matching if problems with ambiguous species names arise.

verbose
has currently no effect.

return
a character vector containing the informatoin that should be extracted into the output. Valid entries are the column names returned by function get_gbif_taxonomy(). See 'Details'.

Details

Taxonomic standardisation is an enormous challenge for biodiversity data management and research. Constant changes in species and higher taxa, refinements of phylogenetic trees and changing attribution to original authors, moving species into other genera or difficulties to place species into the Linean nomenclature results in highly fluctuent taxonomic definitions.

As a consequence, there is not one reference for accepted species names and dependin on the field of resaerch and taxonomic focus other authorities will be employed.

For reasons of simplicity and because of its high coverage of taxa, the function standardize.taxonomy() uses the GBIF Backbone Taxonomy as its reference system and resolves all provided species names to the accepted name according to GBIF (resolving misspellings and synonyms in the process). We invite pull requests to make this function more general and enable a choice of a taxonomic reference.

Examples

pulldata("carabids")

#> The dataset 'carabids' is now available for use! 

dataset1 <- as.traitdata(carabids,
  taxa = "name_correct",
  traits = c("body_length", "antenna_length", "metafemur_length"),
  units = "mm",
  keep = c(datasetID = "source_measurement", measurementRemark = "note"),
  metadata = list(
    bibliographicCitation = attributes(carabids)$citeAs,
    author = "Fons van der Plas",
    license = "http://creativecommons.org/publicdomain/zero/1.0/"
    )
)

#> Input is taken to be a species -- trait matrix. If this is not the case, please provide parameters!

dataset1Std <- standardize.taxonomy(dataset1)

#> 
#> Retrieving data for taxon 'Abax parallelepipedus'

#> 
#> Retrieving data for taxon 'Acupalpus meridianus'

#> 
#> Retrieving data for taxon 'Agonum ericeti'

#> 
#> Retrieving data for taxon 'Agonum fuliginosum'

#> 
#> Retrieving data for taxon 'Agonum gracile'

#> 
#> Retrieving data for taxon 'Agonum marginatum'

#> 
#> Retrieving data for taxon 'Agonum muelleri'

#> 
#> Retrieving data for taxon 'Agonum sexpunctatum'

#> 
#> Retrieving data for taxon 'Agonum viduum'

#> 
#> Retrieving data for taxon 'Amara aenea'

#> 
#> Retrieving data for taxon 'Amara apricaria'

#> 
#> Retrieving data for taxon 'Amara aulica'

#> 
#> Retrieving data for taxon 'Amara bifrons'

#> 
#> Retrieving data for taxon 'Amara brunnea'

#> 
#> Retrieving data for taxon 'Amara communis'

#> 
#> Retrieving data for taxon 'Amara curta'

#> 
#> Retrieving data for taxon 'Amara equestris'

#> 
#> Retrieving data for taxon 'Amara famelica'

#> 
#> Retrieving data for taxon 'Amara familiaris'

#> 
#> Retrieving data for taxon 'Amara fulva'

#> 
#> Retrieving data for taxon 'Amara infima'

#> 
#> Retrieving data for taxon 'Amara lunicollis'

#> 
#> Retrieving data for taxon 'Amara ovata'

#> 
#> Retrieving data for taxon 'Amara plebeja'

#> 
#> Retrieving data for taxon 'Amara quenseli'

#> 
#> Retrieving data for taxon 'Amara similata'

#> 
#> Retrieving data for taxon 'Amara spreta'

#> 
#> Retrieving data for taxon 'Amara tibialis'

#> 
#> Retrieving data for taxon 'Anchomenus dorsalis'

#> 
#> Retrieving data for taxon 'Anisodactylus binotatus'

#> 
#> Retrieving data for taxon 'Asaphidion flavipes'

#> 
#> Retrieving data for taxon 'Badister bullatus'

#> 
#> Retrieving data for taxon 'Bembidion bruxellense'

#> 
#> Retrieving data for taxon 'Bembidion guttula'

#> 
#> Retrieving data for taxon 'Bembidion lampros'

#> 
#> Retrieving data for taxon 'Bembidion nigricorne'

#> 
#> Retrieving data for taxon 'Bembidion properans'

#> 
#> Retrieving data for taxon 'Bembidion quadrimaculatum'

#> 
#> Retrieving data for taxon 'Bembidion tetracolum'

#> 
#> Retrieving data for taxon 'Bradycellus caucasicus'

#> 
#> Retrieving data for taxon 'Bradycellus harpalinus'

#> 
#> Retrieving data for taxon 'Bradycellus ruficollis'

#> 
#> Retrieving data for taxon 'Bradycellus verbasci'

#> 
#> Retrieving data for taxon 'Broscus cephalotes'

#> 
#> Retrieving data for taxon 'Calathus ambiguus'

#> 
#> Retrieving data for taxon 'Calathus cinctus'

#> 
#> Retrieving data for taxon 'Calathus erratus'

#> 
#> Retrieving data for taxon 'Calathus fuscipes'

#> 
#> Retrieving data for taxon 'Calathus melanocephalus'

#> 
#> Retrieving data for taxon 'Calathus micropterus'

#> 
#> Retrieving data for taxon 'Calathus rotundicollis'

#> 
#> Retrieving data for taxon 'Carabus arvensis'

#> 
#> Retrieving data for taxon 'Carabus cancellatus'

#> 
#> Retrieving data for taxon 'Carabus granulatus'

#> 
#> Retrieving data for taxon 'Carabus nemoralis'

#> 
#> Retrieving data for taxon 'Carabus nitens'

#> 
#> Retrieving data for taxon 'Carabus problematicus'

#> 
#> Retrieving data for taxon 'Cicindela campestris'

#> 
#> Retrieving data for taxon 'Cicindela hybrida'

#> 
#> Retrieving data for taxon 'Clivina fossor'

#> 
#> Retrieving data for taxon 'Cychrus caraboides'

#> 
#> Retrieving data for taxon 'Cymindis macularis'

#> 
#> Retrieving data for taxon 'Cymindis vaporariorum'

#> 
#> Retrieving data for taxon 'Demetrias atricapillus'

#> 
#> Retrieving data for taxon 'Dromius angustus'

#> 
#> Retrieving data for taxon 'Dyschirius globosus'

#> 
#> Retrieving data for taxon 'Dyschirius politus'

#> 
#> Retrieving data for taxon 'Harpalus affinis'

#> 
#> Retrieving data for taxon 'Harpalus anxius'

#> 
#> Retrieving data for taxon 'Harpalus distinguendus'

#> 
#> Retrieving data for taxon 'Harpalus laevipes'

#> 
#> Retrieving data for taxon 'Harpalus latus'

#> 
#> Retrieving data for taxon 'Harpalus rubripes'

#> 
#> Retrieving data for taxon 'Harpalus rufipalpis'

#> 
#> Retrieving data for taxon 'Harpalus rufipes'

#> 
#> Retrieving data for taxon 'Harpalus solitaris'

#> 
#> Retrieving data for taxon 'Harpalus tardus'

#> 
#> Retrieving data for taxon 'Harpalus griseus'

#> 
#> Retrieving data for taxon 'Leistus rufomarginatus'

#> 
#> Retrieving data for taxon 'Leistus terminatus'

#> 
#> Retrieving data for taxon 'Limodromus assimilis'

#> 
#> Retrieving data for taxon 'Platynus assimilis'

#> Warning: Synonym provided! Automatically set to accepted species Name!

#> 
#> Retrieving data for taxon 'Loricera pilicornis'

#> 
#> Retrieving data for taxon 'Masoreus wetterhallii'

#> 
#> Retrieving data for taxon 'Miscodera arctica'

#> 
#> Retrieving data for taxon 'Nebria brevicollis'

#> 
#> Retrieving data for taxon 'Nebria salina'

#> 
#> Retrieving data for taxon 'Notiophilus aquaticus'

#> 
#> Retrieving data for taxon 'Notiophilus biguttatus'

#> 
#> Retrieving data for taxon 'Notiophilus germinyi'

#> 
#> Retrieving data for taxon 'Notiophilus rufipes'

#> 
#> Retrieving data for taxon 'Notiophilus substriatus'

#> 
#> Retrieving data for taxon 'Olisthopus rotundatus'

#> 
#> Retrieving data for taxon 'Ophonus rufibarbis'

#> 
#> Retrieving data for taxon 'Oxypselaphus obscurus'

#> 
#> Retrieving data for taxon 'Paradromius linearis'

#> 
#> Retrieving data for taxon 'Patrobus atrorufus'

#> 
#> Retrieving data for taxon 'Philorhizus melanocephalus'

#> 
#> Retrieving data for taxon 'Poecilus cupreus'

#> 
#> Retrieving data for taxon 'Poecilus lepidus'

#> 
#> Retrieving data for taxon 'Poecilus versicolor'

#> 
#> Retrieving data for taxon 'Pterostichus diligens'

#> 
#> Retrieving data for taxon 'Pterostichus melanarius'

#> 
#> Retrieving data for taxon 'Pterostichus minor'

#> 
#> Retrieving data for taxon 'Pterostichus niger'

#> 
#> Retrieving data for taxon 'Pterostichus nigrita'

#> 
#> Retrieving data for taxon 'Pterostichus oblongopunctatus'

#> 
#> Retrieving data for taxon 'Pterostichus rhaeticus'

#> 
#> Retrieving data for taxon 'Pterostichus strenuus'

#> 
#> Retrieving data for taxon 'Pterostichus vernalis'

#> 
#> Retrieving data for taxon 'Stenolophus mixtus'

#> 
#> Retrieving data for taxon 'Stenolophus teutonus'

#> 
#> Retrieving data for taxon 'Stomis pumicatus'

#> 
#> Retrieving data for taxon 'Syntomus foveatus'

#> 
#> Retrieving data for taxon 'Syntomus truncatellus'

#> 
#> Retrieving data for taxon 'Synuchus vivalis'

#> 
#> Retrieving data for taxon 'Trechus obtusus'

#> 
#> Retrieving data for taxon 'Trechus quadristriatus'

#> 
#> Retrieving data for taxon 'Trichocellus cognatus'

#> 
#> Retrieving data for taxon 'Dicheirotrichus cognatus'

#> Warning: Synonym provided! Automatically set to accepted species Name!

#> 
#> Retrieving data for taxon 'Trichocellus placidus'

#> 
#> Retrieving data for taxon 'Dicheirotrichus placidus'

#> Warning: Synonym provided! Automatically set to accepted species Name!