Heteroptera morphometry traits

The data sets comprise morphometric trait data of species that were sampled and measured within the a subproject of the Biodiversity Exploratories Project (Fischer et al. 2010) which focuses on the effect of land use on arthropod community composition and related processes in three regions of Germany (Gossner et al. 2014, Simons et al. 2014, Simons et al. 2015).

Source

https://figshare.com/articles/Data_Paper_Data_Paper/3561936; Creative Commons 0. To the extent possible under law, the authors have waived all copyright and related or neighboring rights to this data.

Value

The dataset heteropteraRaw contains multiple observations of each species (occurence table). The dataset heteroptera is a compiled species-trait matrix.

Details

Morphometric measures of Heteroptera sampled in grasslands across three regions of Germany.

Cite this dataset as:

M. Gossner, Martin; K. Simons, Nadja; Höck, Leonhard; W. Weisser, Wolfgang (2016): Morphometric measures of Heteroptera sampled in grasslands across three regions of Germany. figshare. https://doi.org/10.6084/m9.figshare.c.3307611.v1

See also