Combine trait datasets

Method for function rbind() to append objects of class 'traitdata' to each other.

# S3 method for traitdata
rbind(..., metadata = NULL, datasetID = NULL,
  drop = NULL)

Arguments

...
two or more objects of class traitdata.

metadata
a list of metadata entries which are to be added as dataset-level information.

datasetID
a vector of the same length as number of objects. If NULL (default), object names will be returned as ID.

drop
FALSE by default. If true, columns that are not present in all datasets will be dropped.

Details

Metadata are ideally already included in the datasets as attributes (see ?as.traitdata). The function rbind.traitdata() takes a list of lists as its metadata argument. The outer list must have the same length as the provided objects to combine, with each entry containing objects according to the terms of the Ecological Traitdata Standard (http://ecologicaltraitdata.github.io/ETS/#metadata-vocabulary).

A lookup table for dataset details will be appended as an attribute to the output dataset, linked to each entry via the field datasetID. It can be accessed by calling attributes(<dataset>)$datasets.

Examples

pulldata("carabids")

#> The dataset 'carabids' is now available for use! 

dataset1 <- as.traitdata(carabids,
  taxa = "name_correct",
  traits = c("body_length", "antenna_length", "metafemur_length"),
  units = "mm",
  keep = c(datasetID = "source_measurement", measurementRemarks = "note"),
  metadata = as.metadata(
    bibliographicCitation = attributes(carabids)$citeAs,
    author = "Fons van der Plas",
    license = "http://creativecommons.org/publicdomain/zero/1.0/"
    )
)

#> Input is taken to be a species -- trait matrix. If this is not the case, please provide parameters!

traits1 <- as.thesaurus(
 body_length = as.trait("body_length",
   expectedUnit = "mm", valueType = "numeric",
   identifier = "http://t-sita.cesab.org/BETSI_vizInfo.jsp?trait=Body_length"),
 antenna_length = as.trait("antenna_length",
   expectedUnit = "mm", valueType = "numeric",
   identifier = "http://t-sita.cesab.org/BETSI_vizInfo.jsp?trait=Antenna_length"),
 metafemur_length = as.trait("metafemur_length",
   expectedUnit = "mm", valueType = "numeric",
   identifier = "http://t-sita.cesab.org/BETSI_vizInfo.jsp?trait=Femur_length")
)

dataset1Std <- standardize.traits(dataset1, thesaurus = traits1)

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

# occurrence table: 

pulldata("heteroptera_raw")

#> The dataset 'heteroptera_raw' is now available for use! 

dataset2 <- as.traitdata(heteroptera_raw,
  taxa = "SpeciesID",
  traits = c("Body_length", "Antenna_Seg1", "Antenna_Seg2",
    "Antenna_Seg3", "Antenna_Seg4", "Antenna_Seg5", "Hind.Femur_length"),
  units = "mm",
  keep = c(sex = "Sex", references = "Source", lifeStage = "Wing_development"),
  metadata = as.metadata(
    bibliographicCitation = attributes(heteroptera_raw)$citeAs,
    author = "Martin Gossner",
    license = "http://creativecommons.org/publicdomain/zero/1.0/"
    )
)

#> Input is taken to be an occurrence table/an observation -- trait matrix 
#> (i.e. with individual specimens per row and multiple trait measurements in columns). 
#> If this is not the case, please provide parameters! 

dataset2 <- mutate.traitdata(dataset2,
  antenna_length = Antenna_Seg1 + Antenna_Seg2 + Antenna_Seg3 + Antenna_Seg4 + Antenna_Seg3
  )

traits2 <- as.thesaurus(
 body_length = as.trait("body_length",
   expectedUnit = "mm", valueType = "numeric",
   identifier = "http://t-sita.cesab.org/BETSI_vizInfo.jsp?trait=Body_length"),
 antenna_length = as.trait("antenna_length",
   expectedUnit = "mm", valueType = "numeric",
   identifier = "http://t-sita.cesab.org/BETSI_vizInfo.jsp?trait=Antenna_length"),
 metafemur_length = as.trait("metafemur_length",
   expectedUnit = "mm", valueType = "numeric",
   identifier = "http://t-sita.cesab.org/BETSI_vizInfo.jsp?trait=Femur_length")
)

dataset2Std <- standardize.traits(dataset2,
    thesaurus = traits2,
    rename = c(Body_length = "body_length",
               antenna_length = "antenna_length",
               Hind.Femur_length = "metafemur_length"
               )
    )

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

#> Warning: parse_unit() is deprecated. Please use as_units()

database <- rbind(dataset1Std, dataset2Std)

#> Warning: There seems to be no overlap in taxon names of the provided datasets. 
#>             It is recommended to map 'ScientificNameStd' of each dataset to the same thesaurus or ontology!