Standardize Georeference from Biodiversity Exploratories Plot ID

Adds columns of georeference to trait-data table if measurements relate to specimens from the Biodiversity Exploratories plots or regions.

This function requires valid credentials for the Biodiversity Exploratories Information System (BExIS)!

standardize.exploratories(x, plots = "locationID", user = NULL,
  pswd = NULL, getdata = !is.null(user), fillall = TRUE, ...,
  verbose = NULL, warnings = NULL)

Arguments

x
A traitdata table of class 'traitdata'.

plots
Name of column containing the plot IDs. Must match the Exploratories EP_PlotID scheme (e.g. AEG1, AEG12, HEW21, ...).

user
User name for Biodiversity Exploratories Information System (BExIS, https://www.bexis.uni-jena.de); required for download of exact geolocation.

pswd
password for above request. Will be prompted for if not provided.

fillall
if TRUE (default), the output will contain all terms suggested by the glossary and fill empty columns with NA. This is required for an upload of the data to BExIS.

...
If input is a rawdata table of type species-trait matrix or occurence table (wide table) then provide parameters according to as.traitdata().

Value

std

Examples

# NOT RUN {
moths <- read.service(21247, dec = ",")

dataset1 <- as.traitdata(moths, taxa = "species", traits = c(body_mass =
    "weight", wing_length = "wing_length", wing_width = "wing_width", wing_area =
    "wing_area", wing_loading = "wing_loading"),
  keep = c(locationID = "plot"))

dataset1Std <- standardize.exploratories(dataset1)
# }